requirements.txt
blinker==1.4
CacheControl==0.12.6
cachelib==0.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.7
Deprecated==1.2.12
firebase-admin==5.0.1
Flask==1.1.4
flask-buzz==0.1.15
Flask-Cors==3.0.10
Flask-Mail==0.9.1
flask-praetorian==1.2.0
Flask-SQLAlchemy==2.5.1
gcloud==0.18.3
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-cloud-core==1.7.0
google-cloud-firestore==2.1.3
google-cloud-storage==1.39.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.38.1
httplib2==0.19.1
idna==2.10
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.11.3
jwcrypto==0.9.1
MarkupSafe==2.0.1
msgpack==1.0.2
oauth2client==4.1.3
packaging==20.9
passlib==1.7.4
pendulum==2.1.2
proto-plus==1.18.1
protobuf==3.17.3
psycopg2-binary==2.9.1
py-buzz==1.0.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycrypto==2.6.1
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.18.0
python-jwt==3.3.0
pytz==2021.1
pytzdata==2020.1
requests==2.25.1
requests-toolbelt==0.9.1
rsa==4.7.2
six==1.16.0
SQLAlchemy==1.4.18
sseclient==0.0.27
uritemplate==3.0.1
urllib3==1.26.5
Werkzeug==1.0.1
wrapt==1.12.1
